+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/Bubble/bubbleDynamicPanningData.js
@@ -1,4 +1,4 @@
-import { loadPopup } from '@cerner/terra-graphs-docs/lib/example-datasets/popup.js';
+import { loadBubblePopup } from '@cerner/terra-graphs-docs/lib/example-datasets/popup.js';
 import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 
 const bubbleDynamicPanningData = [{
@@ -6,7 +6,7 @@ const bubbleDynamicPanningData = [{
   label: {
     display: 'Bubble set A',
   },
-  onClick: loadPopup,
+  onClick: loadBubblePopup,
   color: Carbon.helpers.COLORS.PINK,
   values: [{
     x: new Date(2016, 0, 1, 12, 0).toISOString(),
@@ -33,7 +33,7 @@ const bubbleDynamicPanningData = [{
   label: {
     display: 'Bubble set A',
   },
-  onClick: loadPopup,
+  onClick: loadBubblePopup,
   color: Carbon.helpers.COLORS.PINK,
   values: [{
     x: new Date(2016, 0, 1, 12, 0).toISOString(),