+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/Gantt/BasicGanttCharts/LabelTruncation.jsx
@@ -1,0 +1,39 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
+import getGanttDefaultConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttDefault';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(getGanttDefaultConfig('#ganttLabelTruncation'));
+graphConfig.showActionLegend = false;
+graphConfig.dateline = [];
+const dataset = [{
+  key: 'track 1',
+  trackLabel: {
+    display:
+    'Really really long project name that cannot be shown realistically',
+  },
+}];
+
+const GanttLabelTruncationExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="ganttLabelTruncation" />
+    </React.Fragment>
+  );
+};
+
+export default GanttLabelTruncationExample;