+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/General/Legend/LegendItemDisable.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
+import lineTimeseries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiTimeseriesData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineTimeseries('#legendItemDisable'));
+
+const contentData = utils.deepClone(exampleData);
+contentData[0].label.isDisabled = true;
+
+const LegendItemDisableGeneralExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    contentData.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="legendItemDisable" />
+    </React.Fragment>
+  );
+};
+
+export default LegendItemDisableGeneralExample;