+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs-Carbon/LineGraph/Regions/Spline/ValueRegion.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import lineTimesries from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineTimeseries';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/valueRegionData';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Line/LineGraph.module.scss';
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineTimesries('#simpleSplineRegion'));
+const contentData = utils.deepClone(exampleData);
+contentData[0].type = Carbon.helpers.LINE_TYPE.SPLINE;
+
+const SimpleSplineRegionExample = () => {
+  React.useEffect(() => {
+    const graph = Carbon.api.graph(graphConfig);
+    contentData.forEach((data) => {
+      graph.loadContent(Carbon.api.line(data));
+    });
+  }, []);
+  return (
+    <React.Fragment>
+      <div id="tooltip" className="initial-tooltip" />
+      <div id="simpleSplineRegion" />
+    </React.Fragment>
+  );
+};
+
+export default SimpleSplineRegionExample;