+++ packages/terra-graphs/tests/wdio/Graphs/ScatterGraph/scatter-graph-spec.js
@@ -32,4 +32,47 @@ Terra.describeViewports('ScatterGraph ', ['tiny', 'medium', 'large'], () => {
       Terra.validates.screenshot('criticality_simple_scatter', { selector: '.carbon-graph-container' });
     });
   });
+
+  describe('Reflow', () => {
+    describe('simple panning', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/scatter-graph/reflow/simple-panning');
+      });
+      it('validates panning left', () => {
+        panningBtn = $('.button-pan-left');
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_left_3', { selector: '.carbon-graph-container' });
+      });
+      it('validates panning right', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_2', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('simple_panning_right_3', { selector: '.carbon-graph-container' });
+      });
+    });
+
+    describe('dynamic data', () => {
+      let panningBtn = null;
+
+      beforeEach(() => {
+        browser.url('/raw/tests/cerner-terra-graphs-docs/graphs/scatter-graph/reflow/dynamic-data');
+      });
+      it('validates dynaimcally updating of a dataset', () => {
+        panningBtn = $('.button-pan-right');
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_data_1', { selector: '.carbon-graph-container' });
+        panningBtn.click();
+        Terra.validates.screenshot('dynamic_data_2', { selector: '.carbon-graph-container' });
+      });
+    });
+  });
 });