+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -11,15 +11,20 @@
   * Added data objects and examples for Timeline graph.
   * Added data objects and examples for Pie chart.
   * Added data objects and examples for Combination graph.
-  * Added critical and basic example for Timeline graph.
+  * Added critical and basic examples for Timeline graph.
   * Added reflow examples for Line Graph.
-  * Added reflow examples for Paired result graph.
+  * Added reflow examples for Paired Result graph.
   * Added reflow examples for Bar graph.
   * Added reflow examples for Bubble Graph.
   * Added reflow examples for Gantt chart.
   * Added reflow examples for Timeline graph.
+  * Added reflow examples for Scatter graph. 
 
 * Changed 
   * Moved core documentation to terra-dev-site.
+  * Updated Bubble graph documentation and examples to be carbon centric.
+  * Updated Line graph documentation and examples to be carbon-centric.
+  * Updated Combination graph documentation and examples to be carbon-centric.
+  * Updated Gantt chart documentation and examples to be carbon-centric.
+  * Updated general graph documentation and examples to be carbon-centric.
   * Updated Paired Result documentation and examples to be carbon-centric.
-